removing misplaced files and restructuring for TVO scaffolding implementation
