+++ src/js/components/MenuPanel.js
