+++ src/js/components/Progress.js
