+++ test/lesson/assets/gltf/donutv4.gltf
