+++ test/lesson/assets/gltf/trashes.gltf
