+++ test/lesson/index.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>3D Object Scene Viewer</title>
+</head>
+
+<body>
+    <h1>test 2 </h1>
+    <div id="root"></div>
+    <!-- <div class="holder"> -->
+    <div id="scene-viewer-holder" data-config="./data/sv-1.json">
+    </div>
+    <div id="scene-viewer-holder" data-config="./data/sv-2.json">
+    </div>
+
+    <!-- </div> -->
+    <!-- </div> -->
+    <script src="../../dist/main.js"></script>
+</body>
+
+</html>
